# src/dorc-web/src/apis/dorc-api/models/DatabaseApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface DatabaseApiModel
 */
export interface DatabaseApiModel {
    /**
     * @type {Array<string>}
     * @memberof DatabaseApiModel
     */
    EnvironmentNames?: Array<string> | null;
    /**
     * @type {boolean}
     * @memberof DatabaseApiModel
     */
    UserEditable?: boolean;
    /**
     * @type {string}
     * @memberof DatabaseApiModel
     */
    Name?: string | null;
    /**
     * @type {string}
     * @memberof DatabaseApiModel
     */
    Tags?: string | null;
    /**
     * @type {string}
     * @memberof DatabaseApiModel
     */
    ServerName?: string | null;
    /**
     * @type {string}
     * @memberof DatabaseApiModel
     */
    AdGroup?: string | null;
    /**
     * @type {number}
     * @memberof DatabaseApiModel
     */
    Id?: number;
    /**
     * @type {string}
     * @memberof DatabaseApiModel
     */
    ArrayName?: string | null;
}