# src/dorc-web/src/apis/dorc-api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Dorc.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/RefData/{id}": {
      "get": {
        "tags": [
          "RefData"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefDataApiModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefDataApiModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/RefDataDatabases/{id}": {
      "get": {
        "tags": [
          "RefDataDatabases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              }
            }
          }
        }
      }
    },
    "/RefDataDatabases": {
      "get": {
        "tags": [
          "RefDataDatabases"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "server",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseApiModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatabaseApiModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RefDataDatabases"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApiModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApiModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApiModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RefDataDatabases"
        ],
        "parameters": [
          {
            "name": "databaseId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBoolResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RefDataDatabases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApiModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApiModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseApiModel"
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              }
            }
          }
        }
      }
    },
    "/RefDataDatabases/ByTag": {
      "get": {
        "tags": [
          "RefDataDatabases"
        ],
        "parameters": [
          {
            "name": "envName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/RefDataDatabases/ByPage": {
      "put": {
        "tags": [
          "RefDataDatabases"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagedDataOperators"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PagedDataOperators"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PagedDataOperators"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDatabaseApiModelListResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDatabaseApiModelListResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiBoolResult": {
        "type": "object",
        "properties": {
          "Result": {
            "type": "boolean"
          },
          "Message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ComponentApiModel": {
        "type": "object",
        "properties": {
          "ComponentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ComponentName": {
            "type": "string",
            "nullable": true
          },
          "ScriptPath": {
            "type": "string",
            "nullable": true
          },
          "NonProdOnly": {
            "type": "boolean"
          },
          "StopOnFailure": {
            "type": "boolean"
          },
          "ParentId": {
            "type": "integer",
            "format": "int32"
          },
          "IsEnabled": {
            "type": "boolean"
          },
          "PSVersion": {
            "type": "string",
            "nullable": true
          },
          "ComponentType": {
            "$ref": "#/components/schemas/ComponentType"
          },
          "TerraformSourceType": {
            "$ref": "#/components/schemas/TerraformSourceType"
          },
          "TerraformGitBranch": {
            "type": "string",
            "nullable": true
          },
          "TerraformSubPath": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ComponentType": {
        "enum": [
          "PowerShell",
          "Terraform"
        ],
        "type": "string"
      },
...
      "DatabaseApiModel": {
        "type": "object",
        "properties": {
          "EnvironmentNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "UserEditable": {
            "type": "boolean"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Tags": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "ServerName": {
            "type": "string",
            "nullable": true
          },
          "AdGroup": {
            "type": "string",
            "nullable": true
          },
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "ArrayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetDatabaseApiModelListResponseDto": {
        "type": "object",
        "properties": {
          "CurrentPage": {
            "type": "integer",
            "format": "int32"
          },
          "TotalItems": {
            "type": "integer",
            "format": "int32"
          },
          "TotalPages": {
            "type": "integer",
            "format": "int32"
          },
          "Items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseApiModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PagedDataFilter": {
        "type": "object",
        "properties": {
          "Path": {
            "type": "string",
            "nullable": true
          },
          "FilterValue": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PagedDataOperators": {
        "type": "object",
        "properties": {
          "Filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PagedDataFilter"
            },
            "nullable": true
          },
          "SortOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PagedDataSorting"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PagedDataSorting": {
        "type": "object",
        "properties": {
          "Path": {
            "type": "string",
            "nullable": true
          },
          "Direction": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProjectApiModel": {
        "type": "object",
        "properties": {
          "ProjectId": {
            "type": "integer",
            "format": "int32"
          },
          "ProjectName": {
            "type": "string",
            "nullable": true
          },
          "ProjectDescription": {
            "type": "string",
            "nullable": true
          },
          "ArtefactsUrl": {
            "type": "string",
            "nullable": true
          },
          "ArtefactsSubPaths": {
            "type": "string",
            "nullable": true
          },
          "ArtefactsBuildRegex": {
            "type": "string",
            "nullable": true
          },
          "TerraformGitRepoUrl": {
            "type": "string",
            "nullable": true
          },
          "SourceDatabase": {
            "$ref": "#/components/schemas/DatabaseApiModel"
          },
          "SourceControlType": {
            "$ref": "#/components/schemas/SourceControlType"
          }
        },
        "additionalProperties": false
      },
      "SourceControlType": {
        "enum": [
          "AzureDevOps",
          "GitHub",
          "FileShare"
        ],
        "type": "string"
      },
...
      "RefDataApiModel": {
        "type": "object",
        "properties": {
          "Project": {
            "$ref": "#/components/schemas/ProjectApiModel"
          },
          "Components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentApiModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TerraformSourceType": {
        "enum": [
          "SharedFolder",
          "Git",
          "AzureArtifact",
          "GitHubArtifact"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://identityserver:5200/connect/authorize",
            "tokenUrl": "https://identityserver:5200/connect/token",
            "scopes": {
              "dorc-api-np.manage": "Access to DORC API"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "dorc-api-np.manage"
      ]
    }
  ]
}